requirements.txt
numpy
pandas
openai
yt_dlp
youtube_transcript_api
timeout-decorator
python-dotenv
langchain
langchain-community
langchain-qdrant
qdrant-client
streamlit
markdownify
httpx
fastapi
uvicorn
grpcio
grpcio-tools
protobuf